[breadboard-cli] move index (and debugging) into route.
